% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRoutines.R
\name{addRoutines}
\alias{addRoutines}
\title{Add Routines}
\usage{
addRoutines(routinesToAdd, descriptions = rep("", length(routinesToAdd)),
  directory = getwd())
}
\description{
Add a new routine to your daily plan.
}
