% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishRoutines.R
\name{finishRoutines}
\alias{finishRoutines}
\title{Record all remaining routines}
\usage{
finishRoutines(completed = T, directory = getwd())
}
\description{
If you have many routines and are not actively recording their completion throughout the day, it can be helpful to check all of them as complete or incomplete at the end of the day.  That is the purpose of this function.  It takes your list of routines and records them all as either complete or incomplete (T/F/NA).  Will not overwrite existing routine records for that day (if some were logged).
}
