% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRoutineRecord.R
\name{newRoutineRecord}
\alias{newRoutineRecord}
\title{Set up a routine record}
\usage{
newRoutineRecord(directory = getwd())
}
\description{
This will set up a directory where you may record your routines and whether you've completed them.
}
