% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordRoutines.R
\name{recordRoutines}
\alias{recordRoutines}
\title{Record Completed Routines}
\usage{
recordRoutines(routinesCompleted, completed = T, directory = getwd())
}
\description{
If you've completed an activity (or an activity was left incomplete) and you'd like to record that you've finished it (or left it unfinished), then log it here.
}
