% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRoutines.R
\name{removeRoutines}
\alias{removeRoutines}
\title{Remove a routine you've set}
\usage{
removeRoutines(routinesToRemove, directory = getwd())
}
\description{
If a routine no longer seems manageable or worthwhile, you can remove it from your list of routines.
}
