% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routinesNotCompleted.R
\name{routinesNotCompleted}
\alias{routinesNotCompleted}
\title{Show routines that aren't complete}
\usage{
routinesNotCompleted(directory = getwd())
}
\description{
A function that only returns routines not yet marked as complete today.
}
