% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showKindOfRoutine.R
\name{showKindOfRoutine}
\alias{showKindOfRoutine}
\title{Find all Routines of a Particular Kind}
\usage{
showKindOfRoutine(description, directory = getwd(), exactMatching = F)
}
\description{
If you use the description feature for routines, you may wish to filter by routines that match a specific description.  For instance, to bring up all your morning routines.
}
