% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRoutines.R
\name{showRoutines}
\alias{showRoutines}
\title{Show Routines you have active}
\usage{
showRoutines(directory = getwd())
}
\description{
This function shows all of your active routines as of the last occasion that routines were set.
}
