% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successRate.R
\name{successRate}
\alias{successRate}
\title{What proportion of activities are true out of activities marked T/F?}
\usage{
successRate(directory = getwd())
}
\description{
Returns a single value which is the proportion of Trues among Trues and Falses.
}
