% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsRoutines.R
\name{tsRoutines}
\alias{tsRoutines}
\title{Track your Success over Time}
\usage{
tsRoutines(directory = getwd(), plot = T, ylim = c(0, 1),
  cex = 0.1, xlab = "Time", ylab = "Success Rate", ...)
}
\arguments{
\item{directory}{the directory in which the routines are kept}

\item{plot}{whether to plot the results}

\item{Other}{various graphical parameters}
}
\value{
Returns a data frame with the success rate by day, and plots a graph if plot is set to true.
}
\description{
Aggregates success rate by day, and graphs the results as a time series of sorts if plot is set to TRUE.
}
